Initial Server Upload
Serves requests to forecast.io from the weather app and modifies the
response to make it more manageable.

Makes requests to maps.googleapis to determine the location of the user
based on the zip code in their request

Caches recent data so that others using the app nearby can take
advantage of data that’s already been captured.